requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
idna==3.2
multidict==5.1.0
typing-extensions==3.10.0.0
yarl==1.6.3
voluptuous==0.12.1
bcrypt==3.1.7
sqlalchemy==1.4.22